requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
cssselect==1.3.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.1
gitdb==4.0.12
GitPython==3.1.45
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langdetect==1.0.9
lxml==6.0.0
MarkupSafe==3.0.2
narwhals==1.48.1
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
pyarrow==21.0.0
pydeck==0.9.1
PyPDF2>=3.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
rfc3986==1.5.0
rpds-py==0.26.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
streamlit==1.47.1
tenacity==9.1.2
textblob==0.19.0
tinysegmenter==0.3
tldextract==5.3.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
